Ajout du contexte lors de la recherche d'adresse
